Support .config subdirectory links
The links are at he subdirectories, not the files within those
subdirectories. This whole area is for configuration. I don't expect
there'd be files in here we wouldn't want versioned and shared across
installations.

Following the pattern in this repo, these are dotted in the resulting
area (.config) but not here (just config).